互相评估

+++ src/main/java/com/demo/service/sys/SysService.java
@@ -1,5 +1,7 @@
 package com.demo.service.sys;
 
+import java.math.BigDecimal;
+
 public interface SysService {
-	Integer findId();
+	BigDecimal findId();
 }
